changed the functionality of the Accounting class
